analysis/requirements.txt
aqua-analysis-grpc-python==1.76.0.2.20251220113937+7b029a9779df
aqua-analysis-protocolbuffers-pyi==33.2.0.1.20251220113937+7b029a9779df
aqua-analysis-protocolbuffers-python==33.2.0.1.20251220113937+7b029a9779df
certifi==2025.11.12
charset-normalizer==3.4.4
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-reflection==1.76.0
idna==3.11
importlib_metadata==8.7.0
numpy==2.3.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-grpc==0.60b1
opentelemetry-instrumentation-logging==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandas==2.3.3
protobuf==6.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
types-protobuf==6.32.1.20251210
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
wrapt==1.17.3
zipp==3.23.0
